First question in private chat now works.
Added resource bundle and YesNo options for telegram chat parameter.

+++ src/main/java/org/pokescrying/repository/TrainerInfoRepository.java
@@ -6,5 +6,5 @@ import org.pokescrying.data.TrainerInfo;
 import org.springframework.data.repository.CrudRepository;
 
 public interface TrainerInfoRepository extends CrudRepository<TrainerInfo, Long> {
-	Optional<TrainerInfo> findByTelegramId(Integer telegramId);
+	Optional<TrainerInfo> findByTelegramId(Long telegramId);
 }
